fixup! (PC-42928) feat(qpi): Fix navigation intro > questions > intro
