+++ src/features/culturalSurvey/queries/useCulturalSurveyQuestionsQuery.ts
@@ -5,6 +5,7 @@ import { useAuthContext } from 'features/auth/context/AuthContext'
 import { QueryKeys } from 'libs/queryKeys'
 
 const STALE_TIME_CULTURAL_SURVEY_QUESTIONS = 24 * 60 * 60 * 1000
+const GC_TIME_CULTURAL_SURVEY_QUESTIONS = 24 * 60 * 60 * 1000
 
 export const useCulturalSurveyQuestionsQuery = () => {
   const { isLoggedIn } = useAuthContext()
@@ -13,6 +14,7 @@ export const useCulturalSurveyQuestionsQuery = () => {
     queryKey: [QueryKeys.CULTURAL_SURVEY_QUESTIONS],
     queryFn: () => api.getNativeV1CulturalSurveyQuestions(),
     staleTime: STALE_TIME_CULTURAL_SURVEY_QUESTIONS,
+    gcTime: GC_TIME_CULTURAL_SURVEY_QUESTIONS,
     enabled: isLoggedIn,
     meta: { private: true },
   })
